requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# Vector Database
chromadb>=0.4.0

# RAG Framework
langchain>=0.1.0
langchain-community>=0.0.10

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Email System
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
psycopg2-binary>=2.9.0
MSAL>=1.24.0
requests>=2.31.0
langdetect>=1.0.9

# React Frontend (served by Flask)
# Install Node.js separately for React development
